% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiGPlotFunctions.R
\name{getGeneClassColour}
\alias{getGeneClassColour}
\title{Obtains a colour scheme from its corresponding sample class for plotting purposes}
\usage{
getGeneClassColour(
  dFClassName,
  colourList = c("#1b9e77", "#d95f02", "#7570b3", "#e7298a", "#66a61e")
)
}
\arguments{
\item{dFClassName}{A n x 1 dataframe, providing a list of sample classes}

\item{colourList}{A custom colour palette for plotting purposes}
}
\value{
A n x 1 data frame with mapped colour string to its corresponding sample class
}
\description{
Obtains a colour scheme from its corresponding sample class for plotting purposes
}
