% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiGPlotFunctions.R
\name{layoutEpigeneticEV}
\alias{layoutEpigeneticEV}
\title{Create a layout dataset format for plotting and better data representation}
\usage{
layoutEpigeneticEV(
  dataFrame,
  normalized = FALSE,
  sample.class = c(),
  class.colour = c()
)
}
\arguments{
\item{dataFrame}{The n x 4 data frame loaded from csv/rda using the functions above}

\item{normalized}{Whether or not to perform normalization on expression values}

\item{sample.class}{Chooses and subset a specific sample class from the data frame and use it for layout}

\item{class.colour}{User can specify a list/vector custom colour palette of length equals to number of given sample classes}
}
\value{
Returns a layout containing 4 columns of: expression value, quantile over all genes, colour, sample class with their row names assigned as their sample genes
}
\description{
Cleans up the parsed csv/rda file data from obtained dataset, and returns a layout that can be used for plotting
}
\examples{
\dontrun{
rawData <- EpiGPlot::loadEpigeneticData("data/NO66_HUMAN.rda")
gLayout <- EpiGPlot::layoutEpigeneticEV(rawData,
                                       sample.class=c("cell_line", "tissue"),
                                       class.colour=c("red", "blue"))
nrow(gLayout) # 390
unique(gLayout$Colour) # "red", "blue"
}
}
