% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NO66_HUMAN}
\alias{NO66_HUMAN}
\title{The expression values of NO66_HUMAN Gene amongst other information of sample genes}
\format{
A data frame with 889 rows of 4 variables:
\describe{
\item{Sample Class}{Sample Class, the sample gene's respective classification}
\item{Sample}{Sample name of the gene}
\item{Expression Value}{The expression values of the associated sample gene}
\item{Quantile over all genes}{The quantile percentage over all genes}
}
}
\source{
\url{https://epifactors.autosome.ru/genes}
}
\usage{
NO66_HUMAN
}
\description{
A dataset containing the expression values of sample genes as well as their
respective sample class and quantile over all genes information
}
\keyword{datasets}
