% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiGPlotFunctions.R
\name{plotEpigeneticEV}
\alias{plotEpigeneticEV}
\title{Plots the the expression values of a epigenetic gene.}
\usage{
plotEpigeneticEV(
  layout,
  normalized = FALSE,
  colour.legend = TRUE,
  colour.grey = FALSE,
  size.arrow = 1,
  size.pred = 1,
  labels.which = NULL,
  labels.top = 3,
  labels.size = 3,
  sample.class = NULL,
  sample.pred = NULL
)
}
\arguments{
\item{layout}{An n by 4 data frame containing necessary information of epigenetic genes information, usually loaded by the package's given function; see parseEpigeneticData and loadEpigeneticData}

\item{normalized}{Whether the dataset "layout" expression value is normalized.}

\item{colour.legend}{Whether legend on the plot should be dispalyed}

\item{colour.grey}{Whether if greyscale should be applied to all visual details of the plot}

\item{size.arrow}{The size of the line segment if user chooses to display top expression values}

\item{size.pred}{The size of linear regression model's line segment if user chooses to display predictions on sample class}

\item{labels.which}{User can specify which sample to be displayed on the plot through a list/vector}

\item{labels.top}{User can specify how many highest expression value samples to be displayed}

\item{labels.size}{The font size of sample names}

\item{sample.class}{User can specify which sample class to display}

\item{sample.pred}{User can specify which sample class to perform linear regression on, and display their respective line segments}
}
\value{
Returns a plot that user can assign from its return value.
}
\description{
Plots the expression vs quantile data in a scatterplot, linear regression can be visualized for user-specified sample class amongst other user-specified parameters
}
