% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiGPlotFunctions.R
\name{pScaleRange}
\alias{pScaleRange}
\title{Obtains a scale range from 0 to 1 for normalized value plotting purposes}
\usage{
pScaleRange(dFExp)
}
\arguments{
\item{dFExp}{a single column of data frame that provides expression values of sample genes}
}
\value{
Returns the vectorized operated scaled values over the entire column
}
\description{
Obtains a scale range from 0 to 1 for normalized value plotting purposes
}
