% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEpiGPlot.R
\name{runEpiGPlot}
\alias{runEpiGPlot}
\title{Launch Shiny App For Package EpiGPlot}
\usage{
runEpiGPlot()
}
\value{
No return value but open up a shiny page.
}
\description{
A function that launches the shiny app for this package.
The shiny app allows us to easily import epigenetic factor data sets,
processing of data set if needed, and easily control plotting parameters
as well as plot features.
}
\examples{
\dontrun{
runEpiGPlot()
}

}
\references{
Medvedeva, Y. A., Lennartsson, A., Ehsani, R., Kulakovskiy, I. V., Vorontsov, I. E., Panahandeh, P., Khimulya, G., Kasukawa, T., & Drabløs, F. (2015). Epifactors: A comprehensive database of human epigenetic factors and complexes. Database, 2015.
\href{http://database.oxfordjournals.org/content/2015/bav067.full}{Link}

Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
