% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiGPlotFunctions.R
\name{loadEpigeneticData}
\alias{loadEpigeneticData}
\title{Loads epigenetic gene's data from .rda file from the data subdirectory}
\usage{
loadEpigeneticData(rda)
}
\arguments{
\item{rda}{A string or char array containing the file name in the data subdirectory including the .rda suffix}
}
\value{
Returns a data frame containing 4 columns of: Sample class, sample name, normalized expression value, and quantile range.
}
\description{
Loads epigenetic gene's data from .rda file from the data subdirectory
}
