% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiGPlotFunctions.R
\name{parseEpigeneticData}
\alias{parseEpigeneticData}
\title{Parses epigenetic genes expression values from .csv files}
\usage{
parseEpigeneticData(data = "expressions.csv")
}
\arguments{
\item{data}{A file name for the epigenetic gene expression values file}
}
\value{
A data frame containing 4 columns of: Sample class, sample name, normalized expression value, and quantile range.
}
\description{
Requires data to be in data folder, and file should contain 4 columns, separated by tabs.
}
