% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiGPlotFunctions.R
\name{parseEpigeneticData}
\alias{parseEpigeneticData}
\title{Parses epigenetic gene's expression values from .csv files}
\usage{
parseEpigeneticData(data)
}
\arguments{
\item{data}{A file name for the epigenetic gene expression values file including the .csv suffix}
}
\value{
Returns a data frame containing 4 columns of: Sample class, sample name, normalized expression value, and quantile range.
}
\description{
Requires data to be in inst/extdata folder, and file should contain 4 columns, separated by tabs.
}
